Add API for logging using custom winston transport and buffer
Implemented a custom AppTransport that is used for handling log
messages in the actual application. Log entries are routed to a
buffer, which is read and cleared by logActions at a certain interval.
The logActions dispatch log entries, which are in turn rendered by
the LogViewer.

Using a buffer between the AppTransport and actions to avoid slowing
down the application (lots of render calls) when the logger is
invoked very frequently.