Fix ESLint error from eslint-config-airbnb
The airbnb ESLint configuration started throwing error "Configuration
for rule no-use-before-define is invalid". Seems that something has
changed in the airbnb configuration. Not sure how, because all of our
eslint dependencies are pinned.

Followed the instructions in the eslint-config-airbnb documentation
to update to the latest compatible versions. This seems to have fixed
the issue.