Restructure project to support multiple windows
Up until now, we have only had one window. We have the splash screen
of course, but that is just a static page with an image. Now we need
multiple windows that actually renders some dynamic content. We want
to have a loader window that loads and manages plugins, plus the
actual plugin windows.

As these will be separate windows, I propose that we split into two
web applications:
- lib/windows/loader
- lib/windows/plugin

These can share the same API and common code/components from lib,
but are compiled into separate bundles by Webpack.

The loader (when implemented) will be opened first. From here the
user will select a plugin. The loader can then invoke an IPC call
that opens a new plugin window with the name of the plugin as an
URL parameter.