Move NavMenu to common components directory to make it reusable
We are going to use the NavMenu in both the plugin and loader
windows. Moving NavMenu and NavMenuItem to the common components
directory, so that we can use them in both windows.

Also removing the possibility to decorate NavMenuItem. I want
to avoid having references to plugin decoration in the common
components, as they will be used outside of the plugin window. I
do not think decorating each menu item will be a common use case,
and it is still possible to decorate the NavMenu component if
necessary.