Refer to pc-ble-driver-js master during development
Changing to depend on the master branch of pc-ble-driver-js while
we are working on nRF Connect 2.0. This branch contains some new
features, such as getting firmware path/content.

The pc-ble-driver-js master branch will be tagged with a new version
before release, so we should update this dependency when that time
comes.