Remove decoration of reducers
Up until now it has been possible to decorate reducers. However,
this is not used by any of our current apps. We believe that this
is not needed. Removing it to make the API more simple, and to
avoid having to maintain an unnecessary feature. Can consider
re-adding it later if someone asks for it.