Bundle as much as we can with webpack to reduce loading time
Up until now, we have added all dependencies as externals. This is
convenient as we do not have to manually specify which libs should
be external. However, it comes with a performance penalty.

When opening the browser window, all require's of the web page are
processed. If the application needs to load required modules from
node_modules, then it first needs to figure out where the module
exists on the file system, then open the files, and finally parse
them.

If we bundle as much as possible into dist/ at build time, then
the application can skip a lot of the resolving/opening of files
at runtime. In my Windows VM this reduces the loading time of apps
from ~2.2 sec to ~1.5 sec.

It does, however, cause webpack to use more time when building.
We could consider only doing this for production builds if the
longer building time turns out to be annoying.