Specify only the libs we need in production as regular deps
Now that we are bundling most of the libs with webpack, we can
remove some dependencies from the release artifacts. When creating
release artifacts, electron-builder ignores the devDependencies,
so we can move many of our current dependencies to the
devDependencies section.

We must still remember to keep the dependencies used in the main
process as regular dependencies though, as they are not bundled
by webpack.