Fix arrow key navigation in SerialPortSelector
In order for up/down arrow keys to work for the react-bootstrap
DropdownMenu, it needs to have the MenuItem components as direct
children. This means that we cannot wrap the MenuItems in
SerialPortSelectorItem.

Now removing SerialPortSelectorItem and using MenuItem directly
instead. This means that apps cannot decorate
SerialPortSelectorItem anymore, but I do not think that will
be a common use case. Any changes here would probably require
decorating the SerialPortSelector anyway.