Hide "Close" in SerialPortSelector while loading
The SerialPortSelector uses a second to load the ports. When a
port is selected, and the selector is expanded, it would then
first show "Close serial port", and after a short while the
ports would pop up. This is annoying for the user, as the
"Close serial port" item moves position when the ports are
loaded. The user might then mistakenly select a port instead
of closing the port.

To avoid this issue, we now only show the "Close serial port"
item when the ports have been loaded.