Exclude node_modules/pc-nrfjprog-js/nrfjprog
The new pc-nrfjprog-js version comes with tar files and
installers for nRF5x-Command-Line-Tools. When we npm install
pc-nrfjprog-js, these files are included, and added to
node_modules. When creating release artifacts, this adds
an extra ~30 MB to the artifact size, which we do not need.
To avoid this, I am configuring electron-builder to ignore
the given directory.