Avoid performing network requests with Yarn
In our first attempt to make app installation work behind
a proxy (#104), we tried to detect proxy settings with the
Electron network API, and pass it down to Yarn using env
variables. This proved to be unreliable. Electron picks
up proxy settings from the OS in most cases, but we are
not always able to pass it down to the Yarn process. We
can pass the proxy host and port, but detecting and passing
username, password and CA certificates is either unreliable
or in some cases not even possible.

As Electron has much better proxy support then Yarn, we
can take advantage of this by interacting with the npm
registry using the Electron network API instead. The latest
package versions and tarballs are now pulled down by
Electron, and then passed to the Yarn process. This way Yarn
does not have to perform any network requests.

We can now remove all code related to resolving of proxy
settings. By using the Electron network API for all network
requests, we can be sure that it either uses proxy settings
from the OS or from the `--proxy*` flags given when starting
nRF Connect (ref. doc/proxy-settings.md).