Show information dialog if proxy login is cancelled
If the user is behind a proxy that requires authentication, and cancels
the proxy login dialog, then app installation and checking for updates
will fail. Showing some (hopefully) helpful text to the user in this
situation, so that the user can disable checking for updates at startup
and install apps manually instead.