Replace AppProgressDialog with button status text
Now that we are not using yarn for app installation anymore, the apps
install a lot faster. Up until now we have been showing a dialog with
text "Installing pc-nrfconnect-foo. Please wait..." during app
installation, but apps install so fast now that the user will often not
have time to read the text.

Instead of showing the dialog, I am changing the text of the "Install"
button to "Installing..." while app installation is in progress. Also
doing the same for upgrade and remove.