Send network requests in sequence when checking for updates
When the "Check for updates now" button is pressed, we check for both
app updates and core updates. If the user is behind a proxy that
requires authentication, then we cannot send the requests in parallel,
because that would result in HTTP 407 responses for both requests.

We need to send one request first, which receives a HTTP 407, and allow
the user to enter proxy credentials for that request. The next request
will then use the cached credentials.