Verify that serial port is not dead when selected
In some cases, JLink devices stop functioning and become impossible to
open. This typically happens when a JLink device has been connected to
a Windows PC overnight (related to MSD). The only workaround is to power
cycle the device.

When trying to perform operations with nrfjprog when the device is in
this state, it will fail silently with no feedback to the user. At the
moment, the only reliable way to check if a device is actually available
is to try to open it with serialport.