Fix side panel resize issue
When there is some wide content in the main view, the side panel is
pushed to the right and may end up being too narrow or hidden. Applying
this `min-width: 0` flexbox trick that @bencefr used for one of the
apps. This keeps the side panel visible, and adjusts the size of the
main view instead. If the contents of the main view cannot be made
narrower, f.ex. in the BLE app if there are many connections, then a
horizontal scrollbar is shown in the main view.

For more info on this approach, see:
https://css-tricks.com/flexbox-truncated-text/