Use require.resolve to configure webpack loaders
By default, webpack assumes that loaders exist in the root node_modules.
That is not necessarily the case, as the loaders are dependencies of
pc-nrfconnect-devdep, and thus may be in
node_modules/pc-nrfconnect-devdep/node_modules.

The node_modules/pc-nrfconnect-devdep/node_modules path is added to the
NODE_PATH environment variable by pc-nrfconnect-devdep before running
webpack, so we can use `require.resolve('some-loader')` in the webpack
config. This uses the Node.js module system to resolve the paths to the
loaders.