Revert "Always show progress in UpdateProgressDialog"
In the early days, electron-updater did not support progress updates
during auto-update on macOS. We then displayed a spinner instead of a
progress bar on macOS. Progress is now supported on macOS, so the
functionality for showing spinner instead of progress bar was removed.

Now, we are adding support for Linux AppImage auto-updates, and it turns
out that progress updates are not emitted on this platform.
Reintroducing the possibility to show a spinner, so that we can display
it for Linux users.