Make arrow keys work after expanding DeviceSelector
When expanding the DeviceSelector by other means than using the mouse,
it is not possible to use the up/down arrow keys to select a device.
This is because the dropdown button does not have focus.

Now checking if the selector is transitioning into expanded state, and
giving focus to the dropdown button in that case.