Log lister errors to file instead of log viewer
The nrf-device-lister outputs a lot of errors, even in cases where there
is no problems. It tries to get serial numbers from a lot of devices,
and many will fail, and that is to be expected.

Proposing to log the errors with `debug` level instead, so that these
errors end up in the log file. This way they are less intrusive. The
drawback is of course that some users may miss out on important errors.