Disable Cancel button when upgrade is 100% complete
When clicking Cancel while downloading a new version of nRF Connect, and
the download has reached 100%, it does not cancel the operation. The
cancel mechanism in electron-updater is only for cancelling the download
operation. Since the download is complete, cancel does not have any
effect. Disabling the button in this scenario.