Minor simplifications of init routine
The `initAppsDirectory` function originally both intitialized the
`.nrfconnect-apps` directory and downloaded app information (apps.json
and updates.json). The function had error codes to differentiate between
errors related to setting up the directory and errors related to
downloading of app information.

Later, the fetching of app information was moved out from this function
and now happens at the initial render of the launcher window. When doing
this, the JSDoc was not updated, so it refers to an error code
(`APPS_UPDATE_ERROR`) that does not exist anymore.

Updating the JSDoc and simplifying the code slightly.