Rework README
This mostly aims at making the README more approachable for either
people just wanting to run nRF Connect for Desktop or who want to
develop on it. Both have different needs so they are more separated and
each one is more guided, especially on how to get started.

This also corrects the name “nRF Connect” to “nRF Connect for Desktop”.