Fix error dialog in launcher
To accomplish this, the structure of the app states was changed (only
for new apps, not for legacy apps): Before the state contained a slice
“core”, which contained all of the state slices that were handled by
the shared reducers (e.g. appReloadDialog, device, or errorDialog).
This was now changed so that these slices are directly in the main
state, so essentially the middle slice “core” was eliminated and the
slices below were all pulled up on level.