Fix: Launcher could not start in production
This fixes a bug which was introduced in bada7bfc by always depending on
electron-devtools-installer, even though it is only provided when
installing the dev dependencies, so it breaks in production.

So instead we now wrap a try-catch around the `require` and provide the
functionality to install and remove devtools only if that dependency
can be resolved.

Actually, we could always provide the functionality to _remove_ the
devtools, but that would be a strange imbalance.