Weaken compatibility when running from source
When running from source, developers are less interested in app
compatibility checks. So in this case do not warn when launching an app
(the warning on the app icon is still shown, though).

Detecting whether we run from source is done by looking for README.md,
which is not bundled with the app, so it should only be present when
running from source.