Test check for app compatibility
The previous state of checkAppCompatibility.tsx made it a bit hard to
test, because it got the version of the engine from the remote required
config object and the provided version of shared from the project's
package.json. Both were mockable but because there were already loaded
when loading the module it was painful and inflexible. So I opted for
the much easier solution: Enable passing in these things as parameters
and only reading them from the aforementioned places in production.