Model "undecided" explicitly
Previously when a AppCompatibilityChecker returned `undefined` it meant
that it was still undecided whether the app is compatible with the
engine. This is now modeled more explicitly: Checkers never return
undefined but always an object and it contains a property `isDecided`
which is `true` only if a result is found.