Fix: Tailored error for missing source file
The error message when an apps.json for a source is missing on the
server did not show up, because `isResourceNotFound` was called through
a `remote.require`, which strangly lead to the error object not being
filled correctly. Now `isResourceNotFound` is called in the main thread
and thus it works.