Get latest version of uninstalled apps
This commit also has the useful side-effect that it finds et rroneous
entries in apps.json, i.e. if a app is not in a given source but it is
still listed in apps.json, the user will no longer be given the
opportunity to install it (which would then promptly fail).