Remove unused command line options
`--list` and `--install` are not used anymore. `--save-as` was not
implemented anyway.

With this the whole parsing of the command line args by commander, as
well as the code to get the current version of JLink and to install
JLink could also be removed as. The dependencies for sudo-prompt and
regedit also got obsolete.