Position scrollbar directly to the right of apps
Scrollbar spanned the whole height of tab conent, including the
filterbox. This commit fixes the issue by retaining it in a wrapper
which height is set to match the content.

Scrollbar was also positioned to the very right of the window. This
commit positions it closer to the apps to make it more contextualized.