Upgrade electron-store
The new version removes the use of `electron.remote`, which is a plus.

Because we use electron-store also from the main process (in
src/main/apps.js), there is nothing further we have to change. If we
ever remove that usage from the main process in the future, we would
have to add a call to `Store.initRenderer()` in the main process, as
described in the release notes of v7:
https://github.com/sindresorhus/electron-store/releases/tag/v7.0.0