Enhance reliability in China
The app list is not loaded from GitHub any longer but from
developer.nordicsemi.com, which should be easier to reach from China.

This has the consequence though, that we now have to keep
https://developer.nordicsemi.com/.pc-tools/nrfconnect-apps/apps.json in
sync with
https://raw.githubusercontent.com/NordicSemiconductor/pc-nrfconnect-launcher/master/apps.json.
The latter still needs to be maintained as long as we want to also
support older versions of nRF Connect for Desktop.