Fix: Support apps regardless of module system
Before apps needed to export using the CommonJS module system (e.g. by
using `module.exports = … ` or by using `export default …` and the
`add-module-exports` babel plugin).

We want to stop using `add-module-exports` (because it conflicts with
the recent version of electron-store on Windows) but continue to use
`export default …`.

So this change is able to handle apps which are built using the ES
module system (using `export default …` without the `add-module-exports`
babel plugin) by first checking the `default` property. Or falling back
to using the exported object directly for classic apps which use the
CommonJS module system.