Make e2e tests independent of local setup
If a user had set a local filter in the app list or hide the local or
official apps from the app list, some e2e tests would fail.

This change lets the test not use the local setting from the user but
instead the default settings, which prevents these kind of wrong test
failures.