Fix slow npm install
Those slow runs of npm install/npm ci are caused by the combination of
using npm@6 and github:-URLs for some dependencies:
https://twitter.com/colincoller/status/1526369611490508803

When using `github:`, npm (at least in v6, seems to be fixed later)
first tries to use a `git:`-URL and it takes some time until that goes
into a timeout (and is also retried).

So, this removes all uses of github:, which makes `npm ci` run in less
than a minute on my machine (compare to several minutes before.)

GitHub turned off support for git:-URLs for security reasons:
https://github.blog/2021-09-01-improving-git-protocol-security-github/
But it was already switched off in March, so I have no clue why this
escalates now.