Convert remote config to explicit IPC calls
Remote require was previously also using IPC calls in the background but
concealed it. Now it is an explicit call.

Because we only need to get the config once (it does not change later),
it can be cached in the renderer process, so there is done only a single
blocking IPC call now.