Move creating desktop shortcuts to main process
Originally I wanted to remove the `remote.require` from
`desktopShortcutActions.js` but it felt more fitting to move the whole
logic to create desktop shortcuts to the main process. After all, this
is a low level action, not something that should be done in a browser.

So, besides moving `src/launcher/actions/desktopShortcutActions.js` to
`src/main/createDesktopShortcut.js` (and adapting it appropriately),
this meant two IPC channels had to be added: `create-desktop-shortcut`
so that the renderer process can tell the main process to create a
shortcut. And `show-error-dialog`, so that the main process can tell the
renderer process to display an error dialog if something goes wrong.