Convert proxy logins to explicit IPC calls
This is a bit trickier, because when a request indicates (by emitting
the `login` event) that a proxy needs authentication, then the main
process needs to send a message to the renderer process and when the
user has entered the credentials, then these need to be send back to the
main process, where a callback from the original event needs to be
called.

So, when the need for authentication arises, a request ID is created.
With this requst ID the callback is stored in the main process, the
request ID is passed to the renderer process and then back to retrieve
the callback and call it.