Redo logic to install local apps
Expose certain error cases, which we want to handle separately in
following commits. This also slightly enhances edge cases when app
packages are in the folder for local apps and should be unpacked on
startup: E.g. if the file name didn't meet the naming convention,
previously an error was shown but then the launcher did not continue to
start but just hang. Now it shows the error and after confirmation it
continues to start.