Should enforce to use serialport types
I believe it's prettier if we stay to the types that the serialport
library expect us to use, so that potential updates will be more
tidy.

E.g. stay with `{ baudRate: number }`, instead of `baudRate: number`.
Which means that if a type is added to `UpdateOptions`, it will have
minimal impact on this code.