Return void instead of Tuple
Do be consistent with how renderers interact with the serialports
implementation in launcher, we want to return void, and possibly throw,
rather than return a tuple. Thought it got a bit confusing with tuple.

Add proper clean up of the `serialPorts` between each test by clearing
it between each test.