Do not filter out sender on write-notify
This is because this makes it similar for all renderers. With the old
implementation a renderer would write immediately to its terminal, and
then other renderers would be notified.

A problem is that if two renderers write at the same time, the history
would not be equal for the two renderers. They would first write their
own data to terminal, and then receive from the other.

With the new implementation, the data will be written in the order
notified by the IPC, and should therefore be equal for all renderers.

The