Show error if sources are unreadable
If the local `sources.json` is not readable, this will result in a grave
change to how the launcher works: No sources (including their apps)
besides the official and local one will be shown. The user should be
made aware of this situation, so an error in the console is not enough.