Store release notes in file instead of store
This makes the handling more similar to how we handle the icons. The
handling will become even more unified in future commits.

Currently the stored release notes are not yet used as a cache, but this
will also change in a coming commit.

As a side change this requires that we now explicitly initialise the IPC
handler for `electron-store`. This is necessary, because we use
`electron-store` from the renderer process. This was not necessary
before, because it was done implicitly when we created an instance of
`Store` in the main process.