Convert sources from Record to array of objects
Because the latter seems to be the more fitting data structure.

But we for now we keep the old data structure saved on disk in
`source.json` to ease migration. We can already read the new data
structure and later we can switch to also saving the new data structure.