WIP: Working solution for electron@22
Works with the changes that was made in nrf-device-lib-js, together with
the changes in shared in order to clean up the state for devicelib, in
advance to reloading the window.

In this change, I've purposed a pattern that uses IpcRenderer in order
to delegate tasks to the renderer process, and reply when the work is
complete, and only then reload the browser window from the main process.

Since this is a WIP, this commit should be cleaned up before ending
up in the main branch 🤓