Check for minimal app versions
Keep in `minimalRequiredAppVersions.ts` a list of apps where we know,
that the current version of the launcher requires at least a certain app
version. If the installed app is older than that version, then warn
about this. If an app version that is recent enough exists, then inform
the users, that they should upgrade.