In app compatibility checks: Treat ^-ranges as >=
In apps we often specified the required version of the launcher (in the
`engines` property of `package.json`) like `^3.11.1`. This is equivalent
to `>=3.11.1 <4.0.0-0`.

The part `>=3.11.1` was intended but `<4.0.0-0` was not.

We found out, while releasing a new major version of the launcher. Then
several apps claimed, that they are no longer compatible with it. But
this was not our intention. We really meant `>=3.11.1`. But we don't
want to release new versions of those apps just for this small change.

So, this quick fix is to treat something like `^3.11.1` as `>=3.11.1`
for now. With a bit more time we intent to fix the `engines` fields in
all apps. And when all of them are eventually released, then we can also
remove this hack again.