Support withdrawn apps
Handle the case, when an app was once downloadable and installed by
users but is now removed from the server.

Until now, the app then just disappeared from the launcher, even though
it was still installed in the filesystem.

Now, it stays as an installed app, is marked as “not available for
download any more”, and can still be opened as normal. After clicking on
“Uninstall” it completely disappears.