Fix: Strange version number on macOS shortcuts
When creating an app shortcut on macOS, the Finder info for the shortcut
said, that the app has the version 1.2, instead of the actual version of
the app.

This change removes that version number.

Correcting the version number would have been more complicated, because
it would go out of sync when we upgrade the app, unless we implement a
function to also update the version number in the shortcut when
upgrading the app.

Our code does not really use the file Info.plist but instead a copy of
it in template.tar.gz, so both of these files had to be updated.