refactor(ipc): call open:app with AppSpec
As this IPC method will be called more extensively from renderer
processes in order to open new app windows, it's more convenient for the
main process to retrieve the App object from the appName and appSource.

Before this commit, the renderer process would need to retrieve all app
objects, and select the one that matched the appName and appSource. This
commit delegates that responsibility to the main process.