Offer update for changed checksum
To test this, you can e.g. go to an app, build and publish it once to
the test source source with this:
```
npm run build:prod && npm  run  nordic-publish -- -s test
```
Install the app in the launcher.

Then append a newline to `package.json` and build and publish it again:
```
echo >> package.json && npm run build:prod && npm  run  nordic-publish -- -s test
```
Restart the launcher and it will show an update for the app with version
number staying the same.