Replace DIY settings with electron-store
Our `settings` served the same purpose which is already fulfilled by
electron-store, so switch to that.

The `shownSources` are not yet converted because I expect that to be a
little more effort, so I want to do that in a commit on its own.

No migration is done from the old settings to the new code using
electron-store because I think loosing those settings is just a slight
loss of comfort.