Fix: Show Install other versions only if fitting
Only if multiple versions exist the button should be shown. When only a
single version was released, we do not need to show the button.

Occasionally it can also happen, that no versions at all are known. This
case can e.g. happen after a migration from an older launcher and
especially for withdrawn apps.