Make version dropdown scrollable
When there were too many versions to display, the whole window showed a
scrollbar, which would scroll away the dialog.

Now, if there are more than 8 versions available, a scrollbar will be
shown in the dropdown instead.

This also changes the colour of the dropdown to dark, because I found no
easy and not too dirty way to change this.

This also changes the font size of the items in the dropdown list
(before it was bigger, now it is 14px, just like the surrounding text).