Bundle quickstart app with launcher
While packing the app, download
https://developer.nordicsemi.com/.pc-tools/nrfconnect-apps/internal/pc-nrfconnect-quickstart-0.0.1.tgz
and bundle it with the launcher.

When starting the launcher, look for the bundled
pc-nrfconnect-quickstart-0.0.1.tgz. If it is there, install it as a
local app. If there already is a local quickstart app, then remove that
first.

(If there is a ~/.nrfconnect-apps/local/pc-nrfconnect-quickstart/.git
then do not remove and install the local quickstart app, because it
looks like this is development machine.)