Add technical details to error dialog for sources
When a source failed to load, the error message “Unable to retrieve …”
makes an assumption about the possible cause (“… usually caused by a
missing internet connection”) but error reports from users showed us
that we are not sure about this and it might also be caused by other
things. E.g. in https://trello.com/c/RQnktqXb it was caused by something
else.

To understand cases like these better, it is helpful to also get more
technical details, e.g. the exception thrown. But we don’t want to
display raw exceptions in all these cases, because that would not be
user-friendly. So now we display a “Show technical details” button,
which then reveals the text of the exception.