Fix migration: Don't produce broken app info files
When migrating from rather older launcher versions or for a withdrawn
app, updates.json may not contain entries for all apps.

When updates.json doesn't contain an entry for an app, the created app
info file contained `undefined` for the version, because we didn't know
better. But that is not a valid app info file, so we now rather do not
create an app info file in such cases. But this also means, that we have
to filter out apps, which are mentioned in a source.json but for which
no app info file exists yet.