Fix: Multiple requests with authenticated proxies
When a proxy requested authentication in multiple requests at the same
time, only one request was actually continued after the user entered the
credentials (or cancelled the login).

One way to reproduce the problem:
1. Add
   https://developer.nordicsemi.com/.pc-tools/nrfconnect-apps/internal/apps.json
   as a source, so that there are effectively two sources (internal and
   official).
2. Turn on “Check for updates at startup”
3. Restart with a proxy that requires authentication
4. When asked for credentials, enter the correct credentials

Expected result:
- All sources are updated and on the settings page, the button ”Check
  for updates can be clicked”

Actual result:
- Only one source is updated and on the settings page, the button
  ”Checking…” remains disabled

Please note, that when cancelling the login, still a wrong error message
appears. This is a separate issue and will be fixed in a separate PR.