Fix overlaying dialogs with cancelled proxy auth
When cancelling the proxy authentication while getting the latest app
infos at least dialogs were shown:
- The ProxyErrorDialog
- An ErrorDialog for the sources that could not be updated

Now the second dialog is not shown any longer.

When clicking on “Check for updates” in the settings, the
UpdateCheckCompleteDialog was also shown, wrongly stating that all apps
are up to date. This is also not shown any longer