Handle adding existing source more gracefully
Before it was silently accepted, as if it worked correctly, but no new
source was added to the list. Maybe the position of the “added” source
changed in the list of all source.

Now a short message is displayed, why no source was added: No source
added because there already is one with the name “XYZ”.