Fix command line arguments for the bundled app
When the electron app is ran from source the first two command line
arguments are the electron binary and the entry point of the app, so in
that case we need to ignore the first two command line arguments.

When the electron app is bundled, the only the first command line
argument is the binary, so in that case we need to ignore only the first
command line argument.

`process.defaultApp` can be used to distinguish between the two cases:
https://www.electronjs.org/docs/latest/api/process#processdefaultapp-readonly