Use more flexible tsconfig location
The old config only worked if pc-nrfconnect-shared is found top level in
the node_modules folder directly in this project. In some cases it is
found in other places, e.g. when using npm workspaces. The new config is
shorter and uses the npm module resolution mechanism, so the tsconfig
will be found in more places.