Delay showing help on not yet loaded apps
Rational: This case usually only happens while users start nRF Connect
for Desktop the very first time. But even then we do not want to shortly
flash a misleading help text, for potentially the fraction of a second.
So the text is only shown if the app is already running for at least two
seconds.