Move `assertPreparedNrfutilModules` to own file
Motivation: Because it keeps its own state (in `nrfutilSandboxesCache`
fka `existingNrfutilSandboxes`) it seems better to keep the
functionality to assert that the nRF util modules are prepared in a file
on its own.

Also renamed the function because it only prepares the sandboxes on
demand. If they already existed, they are just loaded. And if they were
loaded before, thet are taken from the cache `nrfutilSandboxesCache`.