Rework nrfutilModules
The sandbox cache previously, wrongly would contain entries like
`device-1.2.0,1.3.0` if an app specified that it works with either
version `1.2.0` or `1.3.0` of the `device` module.

Also split up the functionality in nrfutilModules.ts and made use of the
fact that there always must be at least one version specified for a
nrfutil module (which is checked by shared).