Update to shared with new package.json parsing
This version of shared switches to TypeScript for the esbuild-scripts.
Because of this, the `esbuild-renderer` in launcher also needs to be
converted to TypeScript.

We also need to set `skipIgnore` because we execute a script written in
TypeScript which is found in `node_modules/`:
https://www.npmjs.com/package/ts-node#skipping-node_modules

Because the `esbuild-renderer` in shared now can add the react
dependencies to the externals on request, the `esbuild-renderer` in
launcher got rid of the copied code.