fix: change electron version
The newer version of electron was caused by a previous update to electron/remote.
This caused a crash on ubuntu when the cursor entered the application window.

Currently this is just a fix for the crash but it does not stop this from happening again in the future.
A more robust fix would be to have a version of electron required in the launcher.
Given that shared is the one including electron, I have not done this yet to avoid confusion.