Fix: Rare error message on shutdown
When during shutdown some process still tries to send IPC messages to
the launcher window, an exception was thrown, which was presented to
users as an ugly dialog.

To reproduce this, start to install a particular large app (nPM is the
best for this). While the download is still ongoing, quit the launcher.