Test/gh run download (#951)
* Replace github (js) action with gh CLI

It would be beneficial to just use the officially supported GitHub CLI
to download the artifacts, instead of relying on a third-party github
action written in JavaScript.

* Test GitHub CLI to create release

GitHub CLI is already provided by the runner, and is also easy
to use. It also provides option to create release notes using the
Changelog.md file, hence we need to upload it when building.