Fix: Quick Start dialog after J-Link warning
New users, who do not yet have J-Link installed (or a too old version)
and never started the launcher, might have this user flow:

1. Start the launcher and see the Quick Start dialog.
2. From that dialog, try to open the Quick Start app.
3. Get to see the J-Link warning, informing them that they need to
   install a recent enough version of J-Link and then restart nRF
   Connect for Desktop.

Problem: After they have restarted the launcher, they did not see the
Quick Start dialog again. Meaning they first have to understand that and
how they can try to open the Quick Start app again. For the UX it is
nicer if in this case they see the Quick Start dialog again, which is
done by the change.