Fix slow J-Link check on Windows
On Windows running `nrfutil` repeatedly for each app and also before
launching an app made the UX too slow. Especially when launching an app
there was easily a 1-2 second delay.

Now the result of asking for the required J-Link version is memoised per
version of nrfutil-device. For most users this means this will be called
only once or at most a few times and especially not again when launching
an app.

A small consequence by this: Before the users could e.g. install J-Link
and then try to launch an app again (even though the dialog said
otherwise). Now users _must_ really reload the launcher for the launcher
to detect the newly installed J-Link version.