Bump version number
Given the switch from `developer.nordicsemi.com` to
`files.nordicsemi.com`, the version number change for the actual release
will probably be bigger. But for now I just want _some_ version change,
so that for test builds handed out, it is easy to see that they are not
the current release.