Enabling using simple URLs for Artifactory
Now an URL like
https://files.nordicsemi.com/artifactory/swtools/external/ncd/apps/3.7-apps/source.json
can also be used to add a source. Until now it was always necessary to
use the download API URL (e.g.
https://files.nordicsemi.com/ui/api/v1/download?isNativeBrowsing=false&repoKey=swtools&path=external/ncd/apps/3.7-apps/source.json
for the example above).

The shorter URLs are nicer to use and some folks might copy the shorter
URLs from Artifactory (where they are shown as "File URL") and expecting
them to work.